backend/requirements.txt
asgiref==3.6.0
black==22.12.0
click==8.1.3
distlib==0.3.6
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.9.0
mypy-extensions==0.4.3
pathspec==0.10.3
Pillow==9.4.0
platformdirs==2.6.2
PyJWT==2.6.0
pytz==2022.7
sqlparse==0.4.3
tomli==2.0.1
virtualenv==20.17.1